+++ client/src/components/molecules/FormInput/style.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { ifProp, palette, theme } from 'styled-tools';
-import { Input } from 'components';
+import Input from 'components/atoms/Input';
 
 interface Props {
   invalid?: boolean;
